Continue add to cart function with cart window
Made sure the add to cart function worked properly by adding the item to a cart window, showing that item with it's respective information and made sure the cart count reflects the amount of items are in the cart window. Next steps is to total the price and make sure the clear cart button works properly.